Simplify saving
No longer do we need to retrieve the record from the database to
determine whether to do an update or not. No more generated ID. Use the
natural PK of (ESPN ID, Year) to define a league instead. The benefit of
this is that AbstractDAO.persist is better able to know when to do an
update.