Teams!
Basic team functionality. Save 'em to the database only so far, but I do
already have a DAO method for retrieving them.... just need a resource
method.

Made a utility class which will be responsible for building URL's
directly. With the information we're storing, this will be less
expensive then starting from the league main URL and going through
various links to get where we want to go. It'll be useful for ESPN
leagues, at least, given that they put a lot of stuff in the URL
parameters.