Result dates!
A lot of the groundwork for saving results. We'll need an "update"
method that does a minimum of the latest two dates and whatever else
needs doing, and a "full" update that refreshes all results for a given
team. The groundwork is laid for both, though. I'll write the actual
endpoints when I need them for the UI.

Next up is probably Players, since the Stat object needs 'em.